refactor: 스터디 생성파트 에셋 폴더 정리
