+++ app/waiting/page.tsx
@@ -3,14 +3,10 @@
 import { useToast } from '~/components/ui/use-toast'
 import { Toaster } from '~/components/ui/toaster'
 
-import Link from 'next/link'
-import BtnBackIcon from '~/assets/btn_back.svg'
-import BtnMoreVertical from '~/assets/icon_more-vertical.svg'
-import IconBell from '~/assets/icon_bell.svg'
+import IconBell from '~/assets/createStudy/icon_bell.svg'
 import MpProfile from '~/assets/mp_profile.svg'
 import { Badge } from '~/components/ui/badge'
 import { Button } from '~/components/ui/button'
-import { Chip } from '~/components/ui/chip'
 import StudyHeader from '~/components/studycreate/study-header'
 
 export default function WaitingListPage() {
@@ -22,14 +18,6 @@ export default function WaitingListPage() {
       <div className="fixed bottom-[80px] mx-[50px] -translate-x-1/2 transform">
         <Toaster />
       </div>
-
-      {/* <div className="fixed top-4 flex flex-row space-x-[108px] px-3 pt-3">
-        <Link href="approval">
-          <BtnBackIcon />
-        </Link>
-        <h2 className="font-bold">대기중인 요청</h2>
-        <BtnMoreVertical />
-      </div> */}
       <StudyHeader href="approval" />
       <div className="mt-[70px] h-1 w-[375px] border-transparent bg-slate-200"></div>
